2017-09-30 updates.
* Added sound effects for player and enemy shots, shot impacts and craft explosions
* Added music to menu and game scenes
* Added bootstrap loading scene
* Prevented scoring of enemy ships destroyed by Boundary
* Added high score to main menu
* Added player lives and respawning
* Added Game Over message and delayed transfer back to main menu
* Created WebGL build

This represents the end of this learning project.  There are many features that I would add to this game in order to add complexity, variety and (hopefully) enjoyment - if this were intended for any kind of audience - but I feel confident that I could implement those features, so will save that for other projects and concentrate on learning new stuff.